// src/model/categoryElementOutput.ts
/**
 * Shopping API
 * KelkooGroup next generation shopping API. Documentation available at https://developers.kelkoogroup.com/.
 *
 * The version of the OpenAPI document: 4.21.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { GoogleProductCategory } from './googleProductCategory';

export class CategoryElementOutput {
  /**
   * Category ID
   */
  'id'?: number;
  /**
   * Category name
   */
  'name'?: string;
  /**
   * Subcategories
   */
  'categories'?: Array<CategoryElementOutput>;
  'googleProductCategory'?: GoogleProductCategory;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
    {
      name: 'id',
      baseName: 'id',
      type: 'number',
    },
    {
      name: 'name',
      baseName: 'name',
      type: 'string',
    },
    {
      name: 'categories',
      baseName: 'categories',
      type: 'Array<CategoryElementOutput>',
    },
    {
      name: 'googleProductCategory',
      baseName: 'google_product_category',
      type: 'GoogleProductCategory',
    },
  ];

  static getAttributeTypeMap() {
    return CategoryElementOutput.attributeTypeMap;
  }
}