// src/model/categoryTreeOutput.ts
/**
 * Shopping API
 * KelkooGroup next generation shopping API. Documentation available at https://developers.kelkoogroup.com/.
 *
 * The version of the OpenAPI document: 4.21.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CategoryElementOutput } from './categoryElementOutput';

export class CategoryTreeOutput {
  'id': number;
  /**
   * Output format: list or tree
   */
  'format'?: string;
  'category'?: CategoryElementOutput;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
    {
      name: 'id',
      baseName: 'id',
      type: 'number',
    },
    {
      name: 'format',
      baseName: 'format',
      type: 'string',
    },
    {
      name: 'category',
      baseName: 'category',
      type: 'CategoryElementOutput',
    },
  ];

  static getAttributeTypeMap() {
    return CategoryTreeOutput.attributeTypeMap;
  }
}