// src/model/constraintsAtProfileLevel.ts
export class ConstraintsAtProfileLevel {
  'country': string;
  'profileId': number;
  'trackingIds': Array<object>;
  'creationDate': number;
  'isCurrent': boolean;
  'authorityId': number;
  'merchantIds': Array<object>;
  'manualMerchantIds': Array<object>;
  'autoMerchantIds': Array<object>;
  'merchantIdsAreExcluded': boolean;
  'extranetMerchantIds': Array<object>;
  'extranetMerchantIdsAreExcluded': boolean;
  'excludedMerchantGroupIds': Array<object>;
  'categoryIds': Array<object>;
  'categoryIdsAreExcluded': boolean;
  'freeMerchantsAreIncluded': boolean;
  'mobileOffersOnly': boolean;
  'discountedOffersOnly': boolean;
  'deliveryPriceOffersOnly': boolean;
  'minimumCpc': number;
  'applyMobileMultiplierToMinCpc': boolean;
  'excludeAutomodeCPAMerchants': boolean;
  'excludeManualCPAMerchants': boolean;
  'isManualUpdate': boolean;
  'isAutomaticUpdate': boolean;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
    {
      name: 'country',
      baseName: 'country',
      type: 'string',
    },
    {
      name: 'profileId',
      baseName: 'profileId',
      type: 'number',
    },
    {
      name: 'trackingIds',
      baseName: 'trackingIds',
      type: 'Array<object>',
    },
    {
      name: 'creationDate',
      baseName: 'creationDate',
      type: 'number',
    },
    {
      name: 'isCurrent',
      baseName: 'isCurrent',
      type: 'boolean',
    },
    {
      name: 'authorityId',
      baseName: 'authorityId',
      type: 'number',
    },
    {
      name: 'merchantIds',
      baseName: 'merchantIds',
      type: 'Array<object>',
    },
    {
      name: 'manualMerchantIds',
      baseName: 'manualMerchantIds',
      type: 'Array<object>',
    },
    {
      name: 'autoMerchantIds',
      baseName: 'autoMerchantIds',
      type: 'Array<object>',
    },
    {
      name: 'merchantIdsAreExcluded',
      baseName: 'merchantIdsAreExcluded',
      type: 'boolean',
    },
    {
      name: 'extranetMerchantIds',
      baseName: 'extranetMerchantIds',
      type: 'Array<object>',
    },
    {
      name: 'extranetMerchantIdsAreExcluded',
      baseName: 'extranetMerchantIdsAreExcluded',
      type: 'boolean',
    },
    {
      name: 'excludedMerchantGroupIds',
      baseName: 'excludedMerchantGroupIds',
      type: 'Array<object>',
    },
    {
      name: 'categoryIds',
      baseName: 'categoryIds',
      type: 'Array<object>',
    },
    {
      name: 'categoryIdsAreExcluded',
      baseName: 'categoryIdsAreExcluded',
      type: 'boolean',
    },
    {
      name: 'freeMerchantsAreIncluded',
      baseName: 'freeMerchantsAreIncluded',
      type: 'boolean',
    },
    {
      name: 'mobileOffersOnly',
      baseName: 'mobileOffersOnly',
      type: 'boolean',
    },
    {
      name: 'discountedOffersOnly',
      baseName: 'discountedOffersOnly',
      type: 'boolean',
    },
    {
      name: 'deliveryPriceOffersOnly',
      baseName: 'deliveryPriceOffersOnly',
      type: 'boolean',
    },
    {
      name: 'minimumCpc',
      baseName: 'minimumCpc',
      type: 'number',
    },
    {
      name: 'applyMobileMultiplierToMinCpc',
      baseName: 'applyMobileMultiplierToMinCpc',
      type: 'boolean',
    },
    {
      name: 'excludeAutomodeCPAMerchants',
      baseName: 'excludeAutomodeCPAMerchants',
      type: 'boolean',
    },
    {
      name: 'excludeManualCPAMerchants',
      baseName: 'excludeManualCPAMerchants',
      type: 'boolean',
    },
    {
      name: 'isManualUpdate',
      baseName: 'isManualUpdate',
      type: 'boolean',
    },
    {
      name: 'isAutomaticUpdate',
      baseName: 'isAutomaticUpdate',
      type: 'boolean',
    },
  ];

  static getAttributeTypeMap() {
    return ConstraintsAtProfileLevel.attributeTypeMap;
  }
}