// src/model/errorDTOWithMetadata.ts
/**
 * Shopping API
 * KelkooGroup next generation shopping API. Documentation available at https://developers.kelkoogroup.com/.
 *
 * The version of the OpenAPI document: 4.21.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { SearchDebugMetadata } from './searchDebugMetadata';

export class ErrorDTOWithMetadata {
  'error': string;
  'metadata': Array<SearchDebugMetadata>;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
    {
      name: 'error',
      baseName: 'error',
      type: 'string',
    },
    {
      name: 'metadata',
      baseName: 'metadata',
      type: 'Array<SearchDebugMetadata>',
    },
  ];

  static getAttributeTypeMap() {
    return ErrorDTOWithMetadata.attributeTypeMap;
  }
}