// src/model/fieldStats.ts
/**
 * Shopping API
 * KelkooGroup next generation shopping API. Documentation available at https://developers.kelkoogroup.com/.
 *
 * The version of the OpenAPI document: 4.21.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class FieldStats {
  'min'?: object;
  'max'?: object;
  'countDistinct'?: object;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
    {
      name: 'min',
      baseName: 'min',
      type: 'object',
    },
    {
      name: 'max',
      baseName: 'max',
      type: 'object',
    },
    {
      name: 'countDistinct',
      baseName: 'countDistinct',
      type: 'object',
    },
  ];

  static getAttributeTypeMap() {
    return FieldStats.attributeTypeMap;
  }
}