// src/model/flatCategory.ts
/**
 * Shopping API
 * KelkooGroup next generation shopping API. Documentation available at https://developers.kelkoogroup.com/.
 *
 * The version of the OpenAPI document: 4.21.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class FlatCategory {
  /**
   * Category name
   */
  'name'?: string;
  /**
   * Category id
   */
  'id'?: number;
  /**
   * Level in the tree of the current category. Level 1 is the root, Level 2 is the universe, Level > 2 are sub categories.
   */
  'level'?: number;
  'level2'?: string;
  'level3'?: string;
  'level4'?: string;
  'level5'?: string;
  'level6'?: string;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
    {
      name: 'name',
      baseName: 'name',
      type: 'string',
    },
    {
      name: 'id',
      baseName: 'id',
      type: 'number',
    },
    {
      name: 'level',
      baseName: 'level',
      type: 'number',
    },
    {
      name: 'level2',
      baseName: 'level2',
      type: 'string',
    },
    {
      name: 'level3',
      baseName: 'level3',
      type: 'string',
    },
    {
      name: 'level4',
      baseName: 'level4',
      type: 'string',
    },
    {
      name: 'level5',
      baseName: 'level5',
      type: 'string',
    },
    {
      name: 'level6',
      baseName: 'level6',
      type: 'string',
    },
  ];

  static getAttributeTypeMap() {
    return FlatCategory.attributeTypeMap;
  }
}